Move spinbox classes to separate files
- Move ScientificSpinBox from levelswidget.{h,cpp} to
scientificspinox.{h,cpp}
- Move SpinboxWithSlider from levelswidget.{h,cpp} to
spinboxwithslider.{h,cpp}
- Add locale minus sign as acceptable character of ScientificSpinBox
